Create example
First, I wrote a template in Markdown. Markdown uses a plain text format, so you can write it in any old word-processing application; I prefer to use Byword.

Second, I copied and pasted this text into a new TextExpander entry. It looks a little messy, but it's fine as-is. The only things I added were an auto-date-fill-in field at the top and single-line fields for all of the morning questions. This allows me to quickly Tab through and complete each field, without interruption, and without resorting to mousing over to the end of each.

I decided to use the abbreviation <code>;daily<.code>, so when I type that anywhere on my Mac, it will immediately expand this wall of text. Starting off the abbreviation with a semicolon prevents spurious input, and I like to think that the word "daily" tricks my brain into using this, um, daily.

Finally, I open my journaling app of choice, Day One, type my new favorite phrase, `;daily`, and fill in the fields. Press `Return` or click `OK`. There are quite a few neat things about Day One. Like its ability to convert Markdown so it looks pretty.

You may have noticed that I added `#Medical` at the end of this wall of text. In Markdown, a single hashtag generally indicates that that text should be a header. However, Day One actually reads a hashtag followed immediately by a word (i.e., when there's no space between them) as a tag. So all of these entries, which I'm ideally creating every day, are easy to filter and search.